Explanation_Engine/requirements.txt
# backend/requirements.txt
httpx==0.27.0
python-dotenv==1.0.0